refactor: 사용자 정보 객체 수정
- 수정 사항
  - 사용자 정보 객체의 username 정보를 사용자의 이름에서 이메일로 변경
