+++ storeReservation/src/main/java/com/store/reservation/member/security/userDetails/SecurityUser.java
@@ -1,7 +1,10 @@
-package com.store.reservation.member.model;
+package com.store.reservation.member.security.userDetails;
 
 import com.store.reservation.member.domain.MemberInformation;
-import lombok.*;
+import lombok.Builder;
+import lombok.Data;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -13,7 +16,7 @@ import java.util.stream.Collectors;
 @Getter
 @Builder
 @RequiredArgsConstructor
-public class SecurityUser implements UserDetails{
+public class SecurityUser implements UserDetails {
 
     private final MemberInformation memberInformation;
 
@@ -23,9 +26,11 @@ public class SecurityUser implements UserDetails{
                 .map(SimpleGrantedAuthority::new)
                 .collect(Collectors.toList());
     }
-    public Long getId(){
+
+    public Long getId() {
         return this.memberInformation.getId();
     }
+
     @Override
     public String getPassword() {
         return this.memberInformation.getPassword();
@@ -56,7 +61,7 @@ public class SecurityUser implements UserDetails{
         return false;
     }
 
-    public boolean isNotSameUser(Long id){
+    public boolean isNotSameUser(Long id) {
         return !this.getId().equals(id);
     }
 }
